requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Google AI SDK for job application AI features
google-generativeai==0.8.3

# Environment variable management
python-dotenv==1.0.0

# HTTP requests for external APIs
requests==2.31.0

# Email functionality for job applications
smtplib
email